+++ requirements/compile.py
@@ -13,19 +13,6 @@ if __name__ == "__main__":
     # https://docs.djangoproject.com/en/3.0/ref/databases/#mysql-db-api-drivers
     subprocess.run(
         [
-            "python3.5",
-            *common_args,
-            "-P",
-            "Django>=2.2,<2.3",
-            "-P",
-            "mysqlclient>=1.3.13",
-            "-o",
-            "py35-django22.txt",
-        ],
-        check=True,
-    )
-    subprocess.run(
-        [
             "python3.6",
             *common_args,
             "-P",